Code rewrite
Rewritten most of the code of this class. 

Recently I had a web designer (@FedericoLevatti) help me rework both the HTML and CSS code that I was using for the project, which I had borrowed from a WYSIWYG web editor after designing a page template on it, and now the webcode has been much improved. Following are some of the advantages of the new code:

-Its much shorter, easier to read and edit.
-Uses a single CSS style sheet instead of two. Also image files are no longer needed to fill the colors of each element within an item's card, as the CSS code now handles that.
-The resulting web page looks better, more aesthetically pleasing (due to the use of some custom fonts and more CSS styling), and it also adapts to the browser window size.

There's actually more, but overall you can get the idea that it was a big step in every direction.